[refactor] optimize footer
